scripts/requirements.txt
yfinance
supabase
python-dotenv
